Seth, Began Recompartmentalization. Broken commit.
